drivers/net: Turn on XDP features
A summary of the flags being set for various drivers is given below.
Note that XDP_F_REDIRECT_TARGET is a feature that can be turned off
and on at runtime. This means that this flag may be set and unset
under RTNL lock protection by the driver. Hence, READ_ONCE must be
used by code loading the flag value.

Also, this flag is not used for synchronization against the availability
of XDP resources on a device. It is merely a hint, and hence the read
may race with the actual teardown of XDP resources on the device. This
may change in the future, e.g. operations taking a reference on the XDP
resources of the driver, and in turn inhibiting turning off this flag.
However, for now, it can only be used as a hint to check whether device
supports becoming a redirection target.

Turn 'hw-offload' feature flag on for:
 - netronome (nfp)
 - netdevsim.

Turn 'native' and 'zerocopy' features flags on for:
 - intel (i40e, ice, ixgbe)
 - mellanox (mlx5).

Turn 'native' features flags on for:
 - broadcom (bnxt)
 - freescale (dpaa, dpaa2)
 - intel (igb)
 - marvell (mvneta, mvpp2)
 - qlogic (qede)
 - sfc
 - socionext (netsec)
 - ti (cpsw)
 - tap
 - veth
 - xen
 - virtio_net.

Turn 'basic' (tx, pass, aborted and drop) features flags on for:
 - netronome (nfp)
 - amazon (ena)
 - cavium (thunder)
 - mellanox (mlx4)
 - hyperv.

Turn 'tx_lock' feature flag on for:
 - amazon (ena)
 - freescale (dpaa2)
 - intel (igb)
 - marvell (mvneta, mvpp2)
 - qlogic (qede)
 - socionext (netsec)
 - ti (cpsw)
 - tap
 - veth
 - xen

Turn 'redirect_target' feature flag on for:
 - amanzon (ena)
 - broadcom (bnxt)
 - freescale (dpaa, dpaa2)
 - intel (i40e, ice, igb, ixgbe)
 - ti (cpsw)
 - marvell (mvneta, mvpp2)
 - sfc
 - socionext (netsec)
 - qlogic (qede)
 - mellanox (mlx5)
 - tap
 - veth
 - virtio_net
 - xen

Signed-off-by: Marek Majtyka <alardam@gmail.com>
Co-developed-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>
Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>