bpf: Implement cred local storage
This patch adds the ability to stash some data in a struct cred. This
allows one to define their own set of capabilities attached to a subject
(by way of storing them in struct cred) and looking them up later at
security hooks to allow/deny some action. These can be much more fine
grained than existing POSIX capabilities supported by Linux.

Later patches introduce a way to associate a struct cred during the
execution of a BPF_PROG_TYPE_SYSCALL, so that the subjective creds are
overridden while control is in the BPF program. This allows binding
credentials to a BPF program, and passing this as a capability to do
something around to other processes.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>