xdp: add dequeue program type for getting packets from a PIFO
Add a new BPF_PROG_TYPE_DEQUEUE, which will be executed by the device
drivers when they can transmit packets (e.g., on xmit complete).

The idea is that the drivers will do something equivalent to:

max_pkts = MIN(dev->free_txring_space, BULK_SIZE);
for (i = 0; i < max_pkts; i++) {
    struct dequeue_ctx ctx = {
           .egress_ifindex = dev->ifindex,
           .txring_free_items = dev->free_txring_space
    };

    ret = BPF_PROG_RUN(dev->dequeue_prog);
    if (ret || !ctx.dequeued_frame)
       break;

    push_to_txring(ctx.dequeued_frame);
}

ring_tx_doorbell(dev);

This way, the dequeue program can make a decision on which PIFO queue to
pull the packet from (by just keeping references to several PIFO maps,
using direct program references, or using a to-be-defined map-in-map
structure), pick one based on whatever criteria it wants, and return a
frame to the driver. And the driver will do bulking by repeatedly calling
the BPF program, similar to what XDP does on RX today.

This implies that if the BPF dequeue program needs to do expensive lookups
to decide on which queue to dequeue from, it will have to do its own
caching or other mitigation mechanism (we don't do that on behalf of it)
between individual runs.

To actually return the packet, the dequeue program uses a helper which will
put a reference to the packet into the context object, so the caller can
read it from there. This allows us to leverage the existing verifier
reference counting support, so that a packet that is dequeued from a PIFO
map must be either dropped or returned before the dequeue program exits.

For now, the dequeue program can only be run using the bpf_prog_run()
syscall; actually hooking it into the TX path is still TBD. Another
outstanding issue is allowing the dequeue program to access the packet
contents., but this should be doable with a just a bit more surgery to the
verifier.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>