libbpf: Optionally fallback to object BTF for CO-RE candidates
This allows a user to tag a struct flavor and have it be relocated
against its canonical type that is present in the object itself. This
requires skipping over the flavor itself when CO-RE candidates are being
added. This allows us to ship a BPF object that can export a stable
interface as long as it doesn't rename members it has added to the
struct, and be able to extend/reorder struct members.

[
FUTURE: Add BTF_KIND_TAG support to trigger this User must annotate the
struct flavor with __attribute__((btf_tag("core:own_btf"))) which can be
understood by libbpf.
]

Question: Does this make sense to move to static linker phase? Even
though the relo->kind will not be BPF_TYPE_ID_LOCAL, we can detect
BPF_TYPE_ID_TARGET and the tag's presence and then try to find possible
candidates, gracefully handle missing case (which CO-RE at runtime will
patch with invalid insn), otherwise patch the instructions and remove
resolved relocations. It seemed like a fair amount of work so I wanted
to ask before doing it that way...

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>