bpf: cpumap: Enable bulking for generic XDP
Since 11941f8a8536 ("bpf: cpumap: Implement generic cpumap"), we
directly added the skb to the ptr_ring (holding the producer lock) and
woke up the kthread each time. It is possible to instead imitate the
native XDP behavior to enable bulk enqueue. We reuse the bq_enqueue
function, but during flush handle the case where a potential skb may be
dropped on failure to be able to add it to the ptr_ring. This is again
detected by the presence of bit 0 in the pointer representation.

When uninstalling the generic XDP program, we need to take care to call
synchronize_rcu_expedited before decrementing the generic_xdp_needed_key
to disable __cpu_map_flush call. This is because any existing XDP
program inside the netif_receive_skb helpers may still bq_enqueue skb,
and for them the flush needs to be invoked before the function returns.
synchronize_rcu_expedited provides the required barrier as it waits for
existing RCU read section to end, effectively guaranteeing that any
future skbs are not enqueued into the cpumap.

synchronize_rcu_expedited is used since we hold the RTNL lock.

Cc: Jesper Dangaard Brouer <brouer@redhat.com>
Cc: Toke Høiland-Jørgensen <toke@redhat.com>
Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>