bpf: Add XDP_REDIRECT support to XDP for bpf_prog_run()
This adds support for doing real redirects when an XDP program returns
XDP_REDIRECT in bpf_prog_run(). This uses the page pool instance of the
loopback device, allocating a new page for each run repetition, and copying
the data there.

To use this facility the caller has to set the new
BPF_F_TEST_XDP_DO_REDIRECT flag when calling bpf_prog_run(). Enabling this
flag is only allowed if not setting ingress_ifindex in the context object
being passed in, and since it means frames will be redirected somewhere
else, returning the data is not supported either (so data_out and ctx_out
must also be unset).

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>