[KKD] bpf: Enable 64-bit return for bpf_prog_run
BPF ABI always uses 64-bit return value, but so far __bpf_prog_run and
higher level wrappers always truncated the return value to 32-bit.
Future patches introducing a new BPF_PROG_TYPE_DEQUEUE return a
PTR_TO_BTF_ID or NULL from the BPF program to the caller. The verifier
is taught to enforce a successful return of such a referenced
PTR_TO_BTF_ID, explicit release, or the return of a NULL pointer to
indicate absence. To be able to use this returned pointer value, the
bpf_prog_run invocation needs to be able to return a 64-bit value.

This change only modifies the __bpf_prog_run, bpf_prog_run wrapper calls
and prog->bpf_func function pointer (that the JIT image is assigned to).
The rest of the higher level functions still do truncation. A new To
avoid code churn in the whole kernel, we let the compiler handle
truncation normally, and allow new call sites to utilize the 64-bit
return value.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>