bpf: Support attaching task_file, task_vma iter to single tgid
Iterating by directly attaching to single task instead of filtering in
these iterators to find the target task is much faster. We use pidfd
to obtain the task_struct to iterate over (for files and vmas). Next
two commits add selftests and benchmarks showcasing the difference
(based on total tasks in the system) in iteration time.

Currently, pidfd only supports stashing reference to a thread group
leader, but in the future it may also refer to individual threads. For
now, it is unlikely we have a pressing need for that since fdtable and
mm is shared in the POSIX thread model. Whenever pidfd gains such
support, it would be nice to revisit whether we want to support passing
thread pidfd using a new flag or allow that by default.

This change is useful for userspace projects like CRIU, where we are
exploring use of BPF iterators in particular to expedite information
collection conventionally done by parsing procfs virtual files.

The pidfd to attach iterator to will be passed in when attaching the
tracing program. Hence, the task reference will be kept until detach.

Cc: Christian Brauner <brauner@kernel.org>
Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>