bpf: Implement automatic XDP feature detection
There are multiple ways in which an XDP program can return an action,
either directly returning a known constant (such as XDP_DROP, XDP_PASS,
etc.) or returning the return value of bpf_redirect/bpf_redirect_map. In
either of these cases, we have some visibility into the possible
outcomes of the XDP program invocation. Even when it uses bpf_redirect
helpers, all possible actions can be XDP_REDIRECT or the actions encoded
in the lower two bits of flags value.

Hence, this patch does a best effort detection in the verifier. It is
best effort because if we fail to detect any feature being used that the
program depends on at runtime, we want to let the program to be
installed on the netdev instead of returning an error at load time to
preserve backwards compatibility.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>