bpf: Factor out non-fd returning helper from bpf_btf_find_by_name_kind
In future patches, we need a helper that searches all kernel BTFs
(vmlinux and module BTFs), and finds the type denoted by 'name' and
'kind'. Turns out bpf_btf_find_by_name_kind already does the same thing,
but it instead returns a BTF ID and optionally fd (if module BTF).

We extract the core code out into a new helper
btf_find_by_name_kind_all, which returns the BTF ID and BTF pointer in
an out parameter. The reference for the returned BTF pointer is only
bumped if it is a module BTF, this needs to be kept in mind when using
this helper.

The user must release the BTF reference iff btf_is_module is true.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>