bpf: Refactor btf_find_field to take callback
Next commit's field type will not be struct, but pointer, and it will
not be limited to one offset, but multiple ones. Make existing
btf_find_struct_field and btf_find_datasec_var functions amenable to use
for finding BTF ID pointers in map value, by taking a callback and
offloading field specific checks to it.

The alignment, and name are checked before the cb is called, so it is
the last point where we can skip field or return an error before the
next loop iteration happens. This is important, because we'll be
potentially reallocating memory inside cb in next commit, so being able
to do that when everything else is in order is going to be more
convenient.

The name parameter is now optional, and only checked if it is not NULL.

The size must be checked in the callback, because in case of PTR it will
instead point to the underlying BTF ID it is pointing to (or modifiers),
so the check becomes wrong to do outside of callback, and the base type
has to be obtained by removing modifiers.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>