compiler_types: Introduce __bpf_ref annotation for BPF verifier
This annotation uses clang's btf_type_tag attribute to encode whether a
certain parameter or struct member is a referenced object, i.e. the code
using it can assume it's lifetime will be valid without having to raise
its reference count.

Next patch will make use of it to elide refcount bumping in local
storage helpers as an optimization. This also ensures that we always
increment the reference in absence of __bpf_ref, which is the safe case,
but can elide the refcount bump in specific cases where it is guaranteed
to be safe.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>