bpf: Implement local storage helper refcount elision
This commit introduces support in verifier to mark register which are
obtained from function arguments, struct members, etc. annotated using
__bpf_ref (which is "bpf_ref" BTF type tag under the hood) with the
PTR_BPF_REF flag, so that it can pass a hidden parameter to socket, and
inode local storage helpers and allow them to elide refcount bumps in
cases where it is not required (i.e. across the local storage helper
call, the reference is guaranteed to be held).

The next few patches will update the kernel sites where BPF programs are
invoked to manually annotate the safe cases and take advantage of this
feature.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>