bpf: Prepare support for invented kfunc calls
Invented kfunc calls are kfunc calls that are added to the program by
the verifier (for the purposes of inserting a direct call to a kfunc),
this will be used to patch in a direct call to the destructor of a
referenced pointer present in a map. Since all the type information is
known to the verifier at program load time, it can avoid indirect call
and directly fixup the instructions to include a kfunc call insn, that
makes direct call to a destructor function when old pointer in a map
that is refcounted is being replaced by a new one by using a helper.

Future patch will make use of this infrastructure. To enable this, we
modify some rules around the type of offset used for
find_kfunc_desc_btf. Earlier, it was already limited to S16_MAX due to
checks in find_kfunc_desc_btf, but now such checks only happen for calls
present in the program (using add_kfunc_call). For invented calls, we
use the S16_MAX+1 to U16_MAX range, which means we need the code looking
up the offset to always consider it as unsigned.

Thankfully that was already the case internally (since it was
anticipated that we may at some point switch to unsigned offset). The
checks for negative signed offset are limited to kfunc calls present in
the program. Note that this is purely an implementation detail, and we
can support unsigned offset for normal kfunc calls by adjusting the
implementation appropriately. For now, this should be enough as it is
unlikely the program contains as many kfunc calls.

Note that we now have to insert module BTF reference and module
reference separately into kfunc_btf_tab when an invented kfunc call
corresponds to a module BTF and func_id. In this case, we don't have
an fd_array, so we have to generate a custom offset for the invented
call (in the range S16_MAX + 1 to U16_MAX), which we abstract over
and a generation counter with a helper, that checks for overflow
conditions and returns an error.

Finally, these invented calls need not go through the check_kfunc_call
stage, since the verifier ensures a valid argument is passed to them,
and since references they are releasing are not tracked in the program
by the verifier.

Future patches will introduce actual destructor functions that will be
use this facility.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>