bpf: Introduce BPF skiplist map
Introduce a scalable, concurrent skiplist implementation as a new BPF
map type BPF_MAP_TYPE_SKIPLIST. This map type allows the user to create
an ordered set of elements. The comparator to be used for the keys will
be supplied by the user as a BPF program, and will be used to compare
keys when traversing the skiplist. However, the wiring up of comparator
BPF prog is done in a separate change.

... TODO ...

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>