bpf: Add copy_map_value_long for copying to remote percpu region
bpf_long_memcpy is used while copying to remote percpu regions from bpf
syscall and helpers, so that the copy is atomic in word sized chunks.
This might not be possible when you copy kptrs from or to percpu maps,
as the alignment or size may not be multiple of word size.

In this case, only use bpf_long_memcpy when special fields are not
present, otherwise use normal memcpy to copy to disjoint regions.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>