bpf: Check XDP features on attach to network device
If the program has been detected to be using XDP features that are
unavaiable on the target device to which it is being attached by way of
netlink or BPF link attach mode, reject the program attach. Note that
SKB mode of attach supports all possible features program could use,
hence we don't need to perform checks for that mode.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>